refactore the code
